create factory to combine soundxyz & catalog transformer
